+++ docs/sphinx/source/matlab-source/mp_opt_model/mosek_log_callback.m
@@ -1,0 +1,1 @@
+../../../../../lib/mosek_log_callback.m