+++ lib/qps_glpk.m
@@ -89,7 +89,7 @@ function [x, f, eflag, output, lambda] = qps_glpk(H, c, A, l, u, xmin, xmax, x0,
 % See also qps_master, glpk.
 
 %   MP-Opt-Model
-%   Copyright (c) 2010-2024, Power Systems Engineering Research Center (PSERC)
+%   Copyright (c) 2010-2026, Power Systems Engineering Research Center (PSERC)
 %   by Ray Zimmerman, PSERC Cornell
 %
 %   This file is part of MP-Opt-Model.
@@ -195,6 +195,15 @@ else
 end
 glpk_opt.msglev = verbose;
 
+%% handle redirection of console output
+if glpk_opt.msglev
+    %% GLPK is writing to console ...
+    if ~mp.logger.manager('write_to_console') ...   %% and active logger is NOT
+        glpk_opt.msglev = 0;    %% disable GLPK console output
+    end
+    %% GLPK doesn't support logging to a file
+end
+
 %% call the solver
 t0 = tic;
 if isempty(AA)
@@ -203,6 +212,33 @@ if isempty(AA)
     emptyA = true;
 else
     emptyA = false;
+end
+if verbose
+    alg_names = {
+        'primal simplex',
+        'dual simplex',
+        'interior',
+    };
+    if isfield(glpk_opt, 'lpsolver') && glpk_opt.lpsolver == 2
+        alg = 3;
+    else
+        alg = 1;
+        if isfield(glpk_opt, 'dual') && glpk_opt.dual
+            dual = glpk_opt.dual;
+            if have_feature('octave', 'vnum') >= 3.007
+                dual = dual -1;
+            end
+            if dual
+                alg = 2;
+            end
+        end
+    end
+    vn = have_feature('glpk', 'vstr');
+    if isempty(vn)
+        vn = '<unknown>';
+    end
+    mp_printf('GLPK Version %s -- %s %s solver\n', ...
+        vn, alg_names{alg}, 'LP');
 end
 [x, f, errnum, extra] = ...
     glpk(c, AA, bb, xmin, xmax, ctype, vtype, 1, glpk_opt);