+++ lib/qps_mosek.m
@@ -172,6 +172,15 @@ else
     mosek_opt = mosek_options;
 end
 
+%% handle redirection of console output
+if verbose
+    log_to_console = mp.logger.manager('write_to_console');
+    log_file_path = mp.logger.manager('path');
+else
+    log_to_console = true;  %% no MOSEK console output
+    log_file_path = '';
+end
+
 %% set up problem struct for MOSEK
 prob.c = p.c;
 if qp
@@ -261,9 +270,23 @@ if verbose
     mp_printf('MOSEK Version %s -- %s %s solver\n', ...
             vn, alg_names{mosek_opt.MSK_IPAR_OPTIMIZER+1}, lpqp);
 end
-cmd = sprintf('minimize echo(%d)', verbose);
+if log_to_console
+    cmd = sprintf('minimize echo(%d)', verbose);
+else
+    cmd = 'minimize';
+end
 t0 = tic;
-[r, res] = mosekopt(cmd, prob, mosek_opt);
+if isempty(log_file_path)
+    [r, res] = mosekopt(cmd, prob, mosek_opt);
+else
+    log_callback = struct( ...
+        'loghandle', [], ...
+        'log', 'mosek_log_callback' ...
+    );
+    mosek_opt.MSK_IPAR_LOG = verbose;   %% not sure this has any effect
+    outstr = evalc('[r, res] = mosekopt(cmd, prob, mosek_opt, log_callback);');
+    mp_printf(outstr);
+end
 output.runtime = toc(t0);
 
 %%-----  repackage results  -----